+++ src/main/java/heavysnow/heath/dto/goal/GoalIdResponseDto.java
@@ -1,4 +1,4 @@
-package heavysnow.heath.dto;
+package heavysnow.heath.dto.goal;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
