+++ src/main/java/heavysnow/heath/dto/goal/GoalUpdateDto.java
@@ -1,4 +1,4 @@
-package heavysnow.heath.dto;
+package heavysnow.heath.dto.goal;
 
 
 import lombok.AllArgsConstructor;
