+++ src/test/java/heavysnow/heath/service/GoalServiceTest.java
@@ -2,10 +2,8 @@ package heavysnow.heath.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import heavysnow.heath.domain.Goal;
-import heavysnow.heath.domain.Member;
-import heavysnow.heath.dto.GoalCreationDto;
-import heavysnow.heath.dto.GoalIdResponseDto;
-import heavysnow.heath.dto.GoalUpdateDto;
+import heavysnow.heath.dto.goal.GoalCreationDto;
+import heavysnow.heath.dto.goal.GoalUpdateDto;
 import heavysnow.heath.dto.MemberDto;
 import heavysnow.heath.repository.GoalRepository;
 import org.junit.jupiter.api.DisplayName;
@@ -13,9 +11,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
