+++ src/test/java/heavysnow/heath/service/MemberServiceTest.java
@@ -1,7 +1,7 @@
 package heavysnow.heath.service;
 
 import heavysnow.heath.domain.Member;
-import heavysnow.heath.dto.GoalCreationDto;
+import heavysnow.heath.dto.goal.GoalCreationDto;
 import heavysnow.heath.dto.MemberDto;
 import heavysnow.heath.exception.NotFoundException;
 import heavysnow.heath.repository.GoalRepository;
@@ -10,12 +10,8 @@ import jakarta.persistence.EntityManager;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.mockito.internal.matchers.Null;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-
-import java.util.ArrayList;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
